Add snakeviz & speedscope profiling options
Extend the postprocessing benchmarking notebook with snakeviz &
speedscope profiling.  Use a variable to specify which mode; it then
either opens a snakeviz tab or generates a speedscope.json file when
postprocessing is complete.

Note that snakeviz's underlying profile is sampled so not all time is
accounted for and sometimes the call stack looks off. Speedscope
captures the time more precisely.

Use speedscope at speedscope.app