Add container entry point into benchmark script
Extract the interesting benchmark code from the notebook into a
standalone benchmark script. Add an entry point command to the docker
file to execute the script.

Because we're using custom jobs not notebooks, we changed the notebooks
API to custom jobs. This means we need to pass the custom job ID to the
benchmark as a parameter: --custom-job-id

This should be using the "args" parameter … we'll see.