Support .keras file format
Previously we downloaded & extracted. Now we just download. We're
dancing around some hard-coded assumptions such as, the extracted
directory from the archive is the base name (w/o the compression
extension) of the archive file name.

The (only?) caveat of the .keras format is that we need to tell the
loader where to find custom objects used in the model. Apparently, this
model only uses the one `Location2D` layer. Unless/until we use a model
that was (re)trained with different custom objects, this won't change.