Use JSON logging
GCP logging understands the JSON logstash format, not whatever Python
loggers output normally.

This adds a helper function to configure a logger with a new handler
which formats the log to JSON.

This should improve our logs in the GCP log viewer, for example it
should no longer incorrectly identify logs as errors.

Fixes #220