Move GCP JSON formatter to root & remove other handlers
We saw duplicate log entries after we added the GCP handler:
https://github.com/dchaley/deepcell-imaging/issues/279

We saw this once locally, but it went away; then it came back on cloud.
This should "just fix" it.

This PR removes any default handlers from the root logger, and moves the
GCP JSON formatter to the root as well (so that ALL loggers use it, not
just our own).

Fixes #279.