Refactor args object + argparse into helper
Move all helper/description text into the pydantic object models. Then,
build the argparse arguments using this data.

This is not only easier to read / more concise, but also, it makes sure
that our argument object is exactly the same as passed-in arguments.
That helps keep the multi-task batch job the same as single-task jobs.