Add segmentation boot disk based on image size
The segmentation job uses local storage to download the input image and
for intermediate storage. The default (30GB) isn't enough for larger
images. This increases the boot disk size based on the input size.

Unfortunately, although we see the added space in the job json, it
doesn't seem to be reflected in the GCE instance.