Use image name as benchmarking input id
Require passing in the image name. Previously, we guessed it from the
filename using the directory containing the npz file. This is incorrect
for QuPath project structures where the filename itself is the image
name. Now, the job submission script has to decide what to call each
image.

Also fix unit tests– we had forgotten to update them after adding the
disks 😅