Use patched versions of location & plot utils
The DeepCell library loads tensorflow_addons, which isn't released for
tensorflow2.17. We actually only use 1 function.

Patch the module with a copy of the function we use :) Then load the
deepcell modules which eventually load that module.

I wonder; how close are we to removing the deepcell dependency
altogether? Is it worth inlining the 3(right?) functions we have to
remove the deepcell dependency? (We'd still need the toolbox & some
other things)